Changed structure
Added an iframe where the content is loaded and will prompt the user to
print as pdf.

Need to implement a better way to check for changes in uploaded data to
the iframe. Right now using postMsg, probably just props. Will see.

Removed html sanitizer b/c the user case is a file that resembles a
virus with so many scripts inside. It's an inside tool with logs, so
even though its not secure, so is the tool we're using right now.

Probably change structure further. Also need to clean libraries.