css improvements, added page title, help modal has all text
Improved css structure of elements, added a small fade in animation to
the modal.

Fixed a bug in the map function for iframe and video element that would
break if the page structure wasnt like the use case.

Added correct text and images to the modal. Images will probably be
translated and edited a bit soon.

Webpage has a correct title and description. Need to find a suitable
icon