Improved webpage configuration, video parsing
Added icons for the webpage tab and fixed manifest. Removed unused files

All elements are now properly configured so the webapp can be built and
all the imgs, icons... will point to the correct place when uploaded to
the server.

Converted checkboxes into a react component, since we'll probably need
more in the future.

Videos are now replaced with their thumbnail before printing if the
option to convert them into links is unselected (only youtube videos so
far)

Improved structure of the scripts to edit html.

Added or fixed most components documentation.

Tried fixing images being able to be cut in two, but apparently we
cannot with our current document structure.