Added options to remove index pages
Added option to remove details and index pages from the document if its
a supported type of book.

Added option to add a title page if index and details are removed.

Improved structure of the edit details module.

Fixed a bug with promise all using the promise object instead of its
value.