Improved font sizing, added more options to control it (WIP)
Font sizing of elements with a fixed font-size style is now working
properly.

Increased size of header elements, since the base text was getting big
but they werent (need to test this more).
Body now has a better margin for our use case books (it was too big by
default in the default styles).
Will probably move all changes to style the body/generic elements to an
independent module soon, when I've tested more.

Added an element to keep track of independent DOM elements that should
have their font changed (some issues that have arisen with specific
components using classes with fixed size, so this is an attempt to be
able to solve them case by case), so individual problems in a book
can be fixed in a case by case manner.