Added local storing of values, fixed bugs, better error handling
Input values are now stored in localstorage in the browser.

Fixed a bug with referencing old dictionaries instead of making copies
of them in the keyvalueAssign element and its state (kept in the
userEdits array in fileuploader).

Added a "partial success" error state, it shows up when at least one of
the custom rules you try to apply has an invalid selector (basically, if
it's name breaks querySelector). Since the rest of the changes get
applied (even the rest of valid selectors), its a partial success and
shown to the user accordingly.

Improved number input handling and added more tailoring of inputs
depending on necessities (now you can ask for floats, let the number be
negative or not, or ask for a hex number (only in the format # + 6
digits)

Added some better documentation.

Improved css of some elements, specially selectors table and the drag
and drop/fileuploader elements.

Improved structure of the files and filenames (should be clearer what
goes with what, and what file does what)