Index links are now empty, small bug fix for weird width attributes
Links from the index table now are non functional (before they pointed
to invalid documents).
There's a commented functionality to make them point to the current book
url, but since courses change every three months, it doesnt seem worth
it.

When modifying width for elements based on user input in the options
menu, it would not detect width values with the px unit in it if that
value (for ex. "12px") was in the width attribute not set via style.
Note that by default, in the width attribute you dont put units if its
in px (it defaults to px), but someone is bound to do it.

Removed a non working js file (the one that tried to convert a elements
to pdf anchors, since I wasnt able to make it work).