Fixed dependency. Improved modals
There was a dependency alert from the async module used in one of the
libraries. Fixed it so that the version is not vulnerable.

Added diferent modals, improved their user friendliness and added their
corresponding contents.

Had to re-push the changes made for new campus compatibility, and the
new default state.

Removed an unused var in the vimeo library.
Added a small spinner when loading the file.